requirements.txt
open3d>=0.17.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
pyyaml>=6.0
openpyxl>=3.1.0
matplotlib>=3.7.0
